+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Compatibility:
 * Make `File.extname` return `'.'` if the path ends with one (#2192, @tomstuart).
 * Include fractional seconds in `Time#inspect` output (#2194, @tomstuart).
 * Add support for `Integer#[Range]` and `Integer#[start, length]` (#2182, @gogainda).
+* Allow private calls with `self` as an explicit receiver (#2196, @wildmaples).
 
 Performance: